:shirt: Change class name SpawnType to SpawnDirection
