+++ src/main/kotlin/com/github/morinoparty/morinoinsect/catching/condition/TimeCondition.kt
@@ -1,6 +1,6 @@
 package com.github.morinoparty.morinoinsect.catching.condition
 
-import com.github.morinoparty.morinoinsect.catching.insect.SpawnType
+import com.github.morinoparty.morinoinsect.catching.insect.SpawnDirection
 import org.bukkit.block.Block
 import org.bukkit.entity.Player
 
@@ -10,7 +10,7 @@ class TimeCondition(
     override fun check(
         catcher: Player,
         block: Block,
-        spawnType: SpawnType
+        spawnDirection: SpawnDirection
     ): Boolean {
         return catcher.world.isDayTime == state
     }
